Untangled the Horvin/Korvin clusterfuck. In the next version the Horvin will likely become the Zhan Che (?) though as TigerShark suggested.
Added the Merkava I in an effort to clean up the canonicity mess the official sources have created. In one of the TROs it says the Merkava IV (explicitly that number) has been in production since 2294, which is impossible since it mounts gear that wasn't available for more than another century then. By purely looking at the things and some slight attempt at reasoning, the Merkava Target Drone from "Birth of the King" is likely the real Mk I (deprecated at the time of the Mackie test).
The Mk VI can't have appeared much sooner than 2400 due to the LRMs it uses alone.
The Mk VII then correctly appears in 2435 (which makes sense again)