Added the Mackie MSK 7JP, an experimental model with two jump jets.
"With the superb maneuverability of the Wasp with jumpjets Hegemony Leadership felt that merging jumpjet technology with their premier unit, the Mackie, would secure them an advantage over the houses.
Several teams of engineers were tasked with outfitting the 100t Mackie with jumpjets. Initial testing did not go well though. The first attempts had JJs mounted in the Mackie's legs, but the control computer system was
unable to adjust for the weight and movements pilots might command a Mackie to do in flight, both of the prototype models crashed straight into the ground as the Mackies toppled over in flight, head first, one of them killing its pilot in the process.
The other design teams learned a lesson from this and henceforth only experimented with back-mounted jump jets. But the Mackie's weight was just too great, comming down hard from flight as the thrusters gave out, overtaxed and undersized, often called for the
failed prototypes to be dismantled on the field to free the pilots.
The only model that saw a limited production run was the Mackie MSK-7PJ, based on the Mackie 7P. of the 64 ever produced all became known as a mechanic's nightmare, often suffering from slightly bent internal structure after a few jumps,
hatches and maintenance shafts only accessible with blowtorches. A steep price to pay for a very short distance it could jump, it still became the favourite of a few select pilots as it proved valuable enough in city fights and sometimes got them out of
situations that would otherwise have likely been lethal for them.